+++ scripts/testing_package/adhoc.py
@@ -1,4 +1,7 @@
 from vietnamadminunits import parse_address, convert_address
+from vietnamadminunits.pandas import standardize_admin_unit_columns, convert_address_column
+import pandas as pd
 
-print(parse_address('xa binh phuc, huyen thang binh, quang nam', mode='LEGACY'))
-# print(convert_address(''))
+# print(parse_address('02 trường chinh, coopmart, p15 tan binh, hcm', mode='LEGACY'))
+
+print(convert_address('02 trường chinh, coopmart, p15 tan binh, hcm'))