+++ vietnamadminunits/database/main.py
@@ -3,11 +3,13 @@ from pathlib import Path
 
 MODULE_DIR = Path(__file__).parent.parent
 
+
 def query(sql: str):
     '''
     Retrieve administrative unit data from the database.
 
     :param sql: SQL string
+
     :return: Data as a list of JSON-like dictionaries. It is compatible with `pd.DataFrame`.
     '''
     with sqlite3.connect(MODULE_DIR / 'data/dataset.db') as conn:
@@ -17,12 +19,14 @@ def query(sql: str):
         records = [dict(r) for r in result.fetchall()]
         return records
 
+
 def get_data(fields='*', table: str='admin_units', limit: int=None):
     '''
     Retrieve administrative unit data from the database.
 
     :param fields: Column name(s) to retrieve.
-    :param table: Table name, either `'admin_units'` (34 provinces) or `'admin_units_legacy'` (legacy 63 provinces).
+    :param table: Table name, either `'admin_units'` (34-province) or `'admin_units_legacy'` (63-province).
+
     :return: Data as a list of JSON-like dictionaries. It is compatible with `pd.DataFrame`.
     '''
     if isinstance(fields, list):
@@ -32,7 +36,3 @@ def get_data(fields='*', table: str='admin_units', limit: int=None):
         sql += f' LIMIT {limit}'
     records = query(sql)
     return records
-
-
-if __name__ == '__main__':
-    print(get_data(fields='*', table='admin_units'))