+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2025-08-14 (Version 1.0.0)
+- Added 2000 alias keywords for legacy ward.
+- Optimize, clean code, add comments to code.
+- `parse_address()` will keep fully street if the highest level is found.
+
 # 2025-08-06 (Version 0.8.0 - 0.9.0)
 ### [vietnamadminunits/data/parser_legacy.json](vietnamadminunits/data/parser_legacy.json)
 Add more than 700 alias keywords for ward level.